content/base/test/test_bug473162-1.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=473162
-->
<head>
  <title>Test for Bug 473162</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=473162">Mozilla Bug 473162</a>
<p id="display"></p>
<div id="content" style="display: none">
<span id="TEST" class="TEST"></span>  
<span id="test" class="test"></span>  
<span id="Test" class="Test"></span>  
<span id="tesT" class="tesT"></span>  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 473162 **/
var nodes = document.getElementsByClassName("test");
is(nodes.length, 1, "Unexpected length");
is(nodes[0], $("test"), "Unexpected first node");
</script>
</pre>
</body>
</html>
